Changed defailt genesis file to have death turned on with max lifetime being genome length times 20.
Added in a new instruction (div-sex-MS) that allows sexual organims to have
a primitive ability to select their mates -- they will only crossover with
another organism that uses the same label as them following the divide.

+++ source/main/analyze.cc
@@ -641,7 +641,7 @@ double cAnalyze::AnalyzeEntropyGivenParent(cAnalyzeGenotype * genotype,
     
     // Calculate entropy ...
     double this_entropy = 0.0;
-    this_entropy -= (1-mut_rate)*log(1-mut_rate)/log(num_insts);
+    this_entropy -= (1.0-mut_rate)*log(1.0-mut_rate)/log((double) num_insts);
     for (int i = 0; i < num_insts; i ++) {
       if (i == parent_inst) { continue; }
       prob[i] = prob[i] * mut_rate;
